requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langgraph>=0.0.20
langchain-community>=0.0.10

# Data validation and configuration
pydantic>=2.0.0
pyyaml>=6.0

# Terminal output formatting
colorama>=0.4.6

# Environment variables
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.1.0
